chore: initial move of content-related files to content folder
