+++ package.json
@@ -15,9 +15,11 @@
     "@hyvor/hyvor-talk-vue": "^1.0.1",
     "@nuxt/content": "^2.13.0",
     "@pinia/nuxt": "0.11.3",
+    "@types/rss": "^0.0.32",
     "date-fns": "^4.1.0",
+    "nitropack": "^2.13.1",
     "nuxt": "4.2.1",
-    "nuxt-content-assets": "^1.8.2",
+    "nuxt-content-assets": "^1.8.3",
     "perfect-scrollbar": "^1.5.6",
     "pinia": "^3.0.4",
     "remark-reading-time": "^2.0.2",
