--- shared/test.d.ts
@@ -1,7 +1,0 @@
-declare global {
-  interface Test {
-    test: string
-  }
-}
-
-export {}
